Fixed a logic problem in firewall resource handling
Previously the firewall resources were not exported unless the node was tagged
with 'packetfilter'. Split export of firewall resources into their own export
classes, so that they are always exported. This way the target node will be able
to realize the firewall resources from all nodes, instead of just those that
have the packetfilter tag.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>