Worked around a problem with missing DNS A records in bacula::filedaemon
Puppet runs on the Bacula Director node were failing if more than one Filedaemon
node did not have a valid A record in DNS. This was because exported firewall
resources had the node's IP-address in it, and that IP-address would be empty if
the DNS lookup (on puppetmaster) using the node's FQDN had failed. The new
behavior is not to export a firewall resource unless the DNS lookup was
successful.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>