Create catalog backup job in filedaemon code
Creating the catalog backup job used to be handled in the main Bacula Director
configuration file. However, the backup job depended on a properly configured
filedaemon. The backup job was thus moved to filedaemon code to cut down
unnecessary dependencies. Catalog backup job is included only if the node is a
Bacula Director as well as Bacula Filedaemon.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>